% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_adj_matrix.R
\name{CreateAdjMatrix}
\alias{CreateAdjMatrix}
\title{Creates an adjacency matrix from a set of edges.}
\usage{
CreateAdjMatrix(edge.set)
}
\arguments{
\item{edge.set}{A data frame with two columns. Each row is an interaction and each column
is a gene involved in the interaction. Optionally, the data frame can additionally contain
a third column containing a numeric confidence measure for each interaction. Interactions
with missing or negative confidence measures are ignored.}
}
\value{
An adjacency matrix representation of the given edge set.
}
\description{
The first step in the ignition workflow (if you are using your own edge set) is 
to convert the edges into an adjacency matrix. This method provides a simple
framework for doing so.
}
\examples{
data(ignition.example.edges)
adj.mat = CreateAdjMatrix(ignition.example.edges)
}

