% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_predictions.R
\name{GeneratePredictions}
\alias{GeneratePredictions}
\title{Generates network scores for all genes in a network}
\usage{
GeneratePredictions(kernel, pheno.gene.set, autocaps = FALSE,
  pseudo.add = 0.1)
}
\arguments{
\item{kernel}{Required. Produced by \code{\link{CreateKernel}} or user may provide own
gene to gene distance kernel. Row and column names must be set to gene symbols.}

\item{pheno.gene.set}{Required. A vector of previously known phenotype-related genes. WIll
automatically be coerced to uppercase unless overriden (see autocaps param).}

\item{autocaps}{Optional. Defaults to TRUE. Converts all entries in pheno.gene.set to
uppercase to ensure accurate mapping to genes in kernel.}

\item{pseudo.add}{Optional. Defaults to 0.1. Adds a small pseudocount to prevent any gene
from having a percentile of exactly 0 or 1 (qnorm produces an infinite Z score otherwise)}
}
\value{
A data frame containing the computed percentiles and Z scores for each gene.
}
\description{
This method takes a gene to gene distance matrix and computes the summed proximity
of each gene to a set of genes. The set of genes is presumably a collection of known
genes for a phenotype of interest.
}
\examples{
data(ignition.example.edges)
adj.matrix = CreateAdjMatrix(ignition.example.edges)
kernel = CreateKernel(adj.matrix)
known.gene.set = c('B', 'I')
GeneratePredictions(kernel, known.gene.set)
}

