\name{ignition-package}
\alias{ignition-package}
\alias{ignition}
\docType{package}
\title{
\packageTitle{ignition}
}
\description{
\packageDescription{ignition}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{ignition}
\packageIndices{ignition}
}
\author{
\packageAuthor{ignition}
Maintainer: \packageMaintainer{ignition}
}
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
