% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ignition.example.edges}
\alias{ignition.example.edges}
\title{Toy gene interaction dataset for illustration of network diffusion behavior}
\format{A data frame with 12 rows and 2 variables:
\describe{
  \item{gene.one}{first gene involved in the interaction}
  \item{gene.two}{second gene involved in the interaction}
  ...
}}
\usage{
ignition.example.edges
}
\description{
A dataset containing sets of fake gene interactions. Here the genes are simply letters
for simplicity
}
\keyword{datasets}

