% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ignition.example.genetic}
\alias{ignition.example.genetic}
\title{Toy gene-level association dataset}
\format{A data frame with 12 rows and 2 variables:
\describe{
  \item{gene}{gene symbol, in this toy example they are just letters}
  \item{pvals}{example p-values for each gene}
  ...
}}
\usage{
ignition.example.genetic
}
\description{
A dataset containing genes and p-values
}
\keyword{datasets}

