\name{ignition-package}
\alias{ignition-package}
\alias{ignition}
\docType{package}
\title{
\packageTitle{ignition}
}
\description{
\packageDescription{ignition}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{ignition}
\packageIndices{ignition}
}
\author{
\packageAuthor{ignition}
Maintainer: \packageMaintainer{ignition}
}
\keyword{ package }
