\name{ignition-package}
\alias{ignition-package}
\alias{ignition}
\docType{package}
\title{
\packageTitle{ignition}
}
\description{
\packageDescription{ignition}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{ignition}
\packageIndices{ignition}
To be completed
}
\author{
\packageAuthor{ignition}

Maintainer: \packageMaintainer{ignition}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation ~~
~~ directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
